+++ src/Misc/Helpers.cpp
@@ -381,7 +381,7 @@ char* DllLoadAddrStrs() {
   result[0] = '\0';
 
   for (i = 0; i < sizeof(dlls) / sizeof(dlls[0]); ++i) {
-    sprintf_s(lineBuf, sizeof(lineBuf), "%s loaded at: 0x%08x.", dlls[i], GetModuleHandle(dlls[i]));
+    sprintf_s(lineBuf, sizeof(lineBuf), "%s loaded at: 0x%08x.", dlls[i], (uint32_t)(GetModuleHandle(dlls[i])));
     strcat_s(result, strMaxLen, lineBuf);
     if (i != (sizeof(dlls) / sizeof(dlls[0]) - 1)) {
       strcat_s(result, strMaxLen, "\n");
@@ -409,7 +409,7 @@ LONG WINAPI ExceptionHandler(EXCEPTION_POINTERS* ptrs) {
       "\tEIP: 0x%08x, ESP: 0x%08x\n"
       "\tCS: 0x%04x, DS: 0x%04x, ES: 0x%04x, SS: 0x%04x, FS: 0x%04x, GS: 0x%04x\n"
       "\tEAX: 0x%08x, EBX: 0x%08x, ECX: 0x%08x, EDX: 0x%08x, ESI: 0x%08x, EDI: 0x%08x, EBP: 0x%08x, FLG: 0x%08x\n",
-      rec->ExceptionCode, rec->ExceptionAddress, base, ctx->Eip, ctx->Esp, ctx->SegCs, ctx->SegDs, ctx->SegEs, ctx->SegSs, ctx->SegFs, ctx->SegGs, ctx->Eax, ctx->Ebx,
+      rec->ExceptionCode, (uint32_t)rec->ExceptionAddress, base, ctx->Eip, ctx->Esp, ctx->SegCs, ctx->SegDs, ctx->SegEs, ctx->SegSs, ctx->SegFs, ctx->SegGs, ctx->Eax, ctx->Ebx,
       ctx->Ecx, ctx->Edx, ctx->Esi, ctx->Edi, ctx->Ebp, ctx->EFlags);
 
   szString = new char[len + 1];
@@ -420,7 +420,7 @@ LONG WINAPI ExceptionHandler(EXCEPTION_POINTERS* ptrs) {
             "\tEIP: 0x%08x, ESP: 0x%08x\n"
             "\tCS: 0x%04x, DS: 0x%04x, ES: 0x%04x, SS: 0x%04x, FS: 0x%04x, GS: 0x%04x\n"
             "\tEAX: 0x%08x, EBX: 0x%08x, ECX: 0x%08x, EDX: 0x%08x, ESI: 0x%08x, EDI: 0x%08x, EBP: 0x%08x, FLG: 0x%08x\n",
-            rec->ExceptionCode, rec->ExceptionAddress, base, ctx->Eip, ctx->Esp, ctx->SegCs, ctx->SegDs, ctx->SegEs, ctx->SegSs, ctx->SegFs, ctx->SegGs, ctx->Eax,
+            rec->ExceptionCode, (uint32_t)rec->ExceptionAddress, base, ctx->Eip, ctx->Esp, ctx->SegCs, ctx->SegDs, ctx->SegEs, ctx->SegSs, ctx->SegFs, ctx->SegGs, ctx->Eax,
             ctx->Ebx, ctx->Ecx, ctx->Edx, ctx->Esi, ctx->Edi, ctx->Ebp, ctx->EFlags);
 
   dllAddrs = DllLoadAddrStrs();
